requirements.txt
opencv-python~=4.10.0.84
keyboard~=0.13.5
numpy~=2.1.1
PyAutoGUI~=0.9.54
pillow~=10.4.0
PyQt6~=6.7.1
pyinstaller~=6.10.0
psutil~=6.0.0
PyCryptodome~=3.20.0